remove unused method to fetch the genesis and gentx file
