+++ starport/pkg/cosmosutil/genesis.go
@@ -1,39 +1,52 @@
 package cosmosutil
 
 import (
+	"encoding/json"
+	"errors"
 	"os"
-	"path/filepath"
 )
 
-// genesisPath returns the default genesis path into the home dir
-func genesisPath(home string) string {
-	return filepath.Join(home, "config/genesis.json")
+// ChainGenesis represents the stargate genesis file
+type ChainGenesis struct {
+	AppState struct {
+		Auth struct {
+			Accounts []struct {
+				Address string `json:"address"`
+			} `json:"accounts"`
+		} `json:"auth"`
+	} `json:"app_state"`
 }
 
-// getChainGenesis return the chain genesis path
-func getChainGenesis(home string) (ChainGenesis, bool, error) {
-	genesisPath := genesisPath(home)
-	_, err := os.Stat(genesisPath)
-	if os.IsNotExist(err) {
-		return ChainGenesis{}, false, nil
-	} else if err != nil {
-		return ChainGenesis{}, false, err
+// HasAccount check if account exist into the genesis account
+func (g ChainGenesis) HasAccount(address string) bool {
+	for _, account := range g.AppState.Auth.Accounts {
+		if account.Address == address {
+			return true
+		}
 	}
-	net, err := ParseGenesis(genesisPath)
+	return false
+}
+
+// ParseGenesis parse ChainGenesis object from a genesis file
+func ParseGenesis(genesisPath string) (genesis ChainGenesis, err error) {
+	genesisFile, err := os.ReadFile(genesisPath)
 	if err != nil {
-		return ChainGenesis{}, false, err
+		return genesis, errors.New("cannot open genesis file: " + err.Error())
 	}
-	return net, true, nil
+	return genesis, json.Unmarshal(genesisFile, &genesis)
 }
 
 // CheckGenesisContainsAddress returns true if the address exist into the genesis file
-func CheckGenesisContainsAddress(home, addr string) (bool, error) {
-	genesis, exist, err := getChainGenesis(home)
-	if err != nil {
+func CheckGenesisContainsAddress(genesisPath, addr string) (bool, error) {
+	_, err := os.Stat(genesisPath)
+	if os.IsNotExist(err) {
+		return false, nil
+	} else if err != nil {
 		return false, err
 	}
-	if !exist {
-		return false, nil
+	genesis, err := ParseGenesis(genesisPath)
+	if err != nil {
+		return false, err
 	}
 	return genesis.HasAccount(addr), nil
 }
